mai_assistant_telegram_bot/requirements.txt
aio_pika==9.3.1
websockets==11.0.3
pika==1.3.2
python-telegram-bot==20.7
requests==2.31.0
redis==5.0.1